[ENHANCEMENT] Move show all toggle to tooltip header beside pin
